[debian] compile for multiple Android versions at the same time
This uses the same technique used by qtubuntu to compile both -android
and -desktop variant at the same time.

It currently builds for API version 28 only, as API version 24 does not
need this package. To add more versions, edit the version list at the
top of the debian/rules file, add packages in the debian/control, and
finally link the .install files.